collision: merge Collide_TestCollision to trx
