+++ src/libtrx/include/libtrx/game/collision.h
@@ -64,6 +64,8 @@ typedef struct {
 
 int32_t Collide_GetSpheres(const ITEM *item, SPHERE *spheres, bool world_space);
 
+int32_t Collide_TestCollision(ITEM *item, const ITEM *lara_item);
+
 void Collide_GetCollisionInfo(
     COLL_INFO *coll, int32_t x, int32_t y, int32_t z, int16_t room_num,
     int32_t obj_height);
