+++ src/tr1/game/collide.c
@@ -12,42 +12,6 @@
 #include <libtrx/game/matrix.h>
 #include <libtrx/utils.h>
 
-int32_t Collide_TestCollision(ITEM *item, const ITEM *lara_item)
-{
-    SPHERE slist_baddie[34];
-    SPHERE slist_lara[34];
-
-    uint32_t flags = 0;
-    int32_t num1 = Collide_GetSpheres(item, slist_baddie, 1);
-    int32_t num2 = Collide_GetSpheres(lara_item, slist_lara, 1);
-
-    for (int i = 0; i < num1; i++) {
-        SPHERE *ptr1 = &slist_baddie[i];
-        if (ptr1->r <= 0) {
-            continue;
-        }
-        for (int j = 0; j < num2; j++) {
-            SPHERE *ptr2 = &slist_lara[j];
-            if (ptr2->r <= 0) {
-                continue;
-            }
-            int32_t x = ptr2->pos.x - ptr1->pos.x;
-            int32_t y = ptr2->pos.y - ptr1->pos.y;
-            int32_t z = ptr2->pos.z - ptr1->pos.z;
-            int32_t r = ptr2->r + ptr1->r;
-            int32_t d = SQUARE(x) + SQUARE(y) + SQUARE(z);
-            int32_t r2 = SQUARE(r);
-            if (d < r2) {
-                flags |= 1 << i;
-                break;
-            }
-        }
-    }
-
-    item->touch_bits = flags;
-    return flags;
-}
-
 void Collide_GetJointAbsPosition(ITEM *item, XYZ_32 *vec, int32_t joint)
 {
     const OBJECT *const obj = Object_Get(item->object_id);
